--- src/contas/contaCorrente.model.ts
@@ -1,5 +1,0 @@
-import { ContaBancaria } from "../contas/contaBancaria.model";
-
-export class ContaCorrente extends ContaBancaria{
-    private limiteChequeEspecial!: number
-}
